add standard number
- เพิ่มหมายเลขมาตราฐานการผลิต

+++ usermanage/standardcontrol.php
@@ -1,13 +1,13 @@
-<meta charset="utf8">
 <?php
 require_once('../connect/connect.php');
 session_start();
-$standard = $_POST['standard'];
-$idfac = $_SESSION['idfactory_place'];
-$idstd = $_GET['id'];
 
-if(!isset($idstd)){
-	$addstandard = $con->query("INSERT INTO standard (standard, idfactory_place, dateadd) VALUES ('$standard', $idfac, now()) ") or die (mysqli_error($con));
+
+if($_POST['action'] == "insert"){
+	$standard = $_POST['standard'];
+	$standardnumber = $_POST['number'];
+	$idfac = $_SESSION['idfactory_place'];
+	$addstandard = $con->query("INSERT INTO standard (standard, idfactory_place, dateadd, number) VALUES ('$standard', $idfac, now(), '$standardnumber') ") or die (mysqli_error($con));
 	if($addstandard){
 		echo '<script type="text/javascript" >
 			alert("เพิ่มมาตราฐานเรียบร้อย");
@@ -19,7 +19,8 @@ if(!isset($idstd)){
 			window.location.href = "javascript:history.back()";
 			</script>';
 	}
-}else{
+}elseif($_POST['action'] == "delete"){
+	$idstd = $_POST['id'];
 	$deletestandard = $con->query("DELETE FROM standard WHERE idstandard = $idstd") or die (mysqli_error($con));
 	if($deletestandard){
 		echo '<script type="text/javascript" >
